# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a loyalty campaigns application.",
        "title": "Loyalty Campaigns API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7070",
    "basePath": "/api",
    "paths": {
...
        "/api/loyalty/process-transaction": {
            "post": {
                "description": "Process a user transaction and award loyalty points or cashback based on active campaigns",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loyalty"
                ],
                "summary": "Process a transaction and award loyalty points or cashback",
                "parameters": [
                    {
                        "description": "Transaction details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/loyalty_requests.ProcessTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/loyalty/redeem-rewards": {
            "post": {
                "description": "Redeem a user's loyalty points or cashback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loyalty"
                ],
                "summary": "Redeem user rewards",
                "parameters": [
                    {
                        "description": "Redemption details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/loyalty_requests.RedeemRewardsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "loyalty_requests.ProcessTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "branchId",
                "date",
                "merchantId",
                "userId"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "branchId": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "merchantId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "loyalty_requests.RedeemRewardsRequest": {
            "type": "object",
            "required": [
                "amount",
                "merchantId",
                "rewardType",
                "userId"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "merchantId": {
                    "type": "integer"
                },
                "rewardType": {
                    "type": "string",
                    "enum": [
                        "points",
                        "cashback"
                    ]
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}